Tests/LibWeb/Text/expected/HTML/History-pushState-change-query.txt

Good: changing the query via pushState()
Good: threw on changing the filename via pushState()
Good: going back to the original filename
